PP-30 refactor: change game cards
