+++ pointing-poker/src/components/Game/GameCard/GameCard.module.scss
@@ -1,0 +1,30 @@
+@import '../../../layout/styles-config';
+
+.gameCard {
+  position: relative;
+  margin-bottom: 5px;
+  width: 100px;
+  height: 160px;
+  @include flex($direction: column);
+  box-shadow: $card-shadow;
+  border-radius: 5px;
+  font-size: 16px;
+
+  &.active {
+    background-color: #03fc4e5c;
+  }
+}
+
+.cover {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 4;
+  &:hover {
+    cursor: pointer;
+    background-color: #03fc4e5c;
+    opacity: 0.5;
+  }
+}
