+++ pointing-poker/src/components/Game/GameCard/GameCard.test.tsx
@@ -1,12 +1,12 @@
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { GameCard } from '..';
+import { GameCard } from '../GameCard';
 
 describe('GameCard', () => {
-  const handleEditScoreValue = jest.fn();
+  const handleSelectCurrentCard = jest.fn();
 
   test('should render component with current props', () => {
-    const { getByTestId } = render(<GameCard scoreType="sp" scoreValue={8} />);
+    const { getByTestId } = render(<GameCard isCurrent={false} scoreType="sp" scoreValue="8" />);
 
     expect(getByTestId(/topCard/)).toHaveTextContent('8');
     expect(getByTestId(/centerCard/)).toHaveTextContent('sp');
@@ -15,14 +15,14 @@ describe('GameCard', () => {
 
   test('should calls the handleEditScoreValue', () => {
     const { getByRole } = render(
-      <GameCard scoreType="sp" scoreValue={8} handleEditScoreValue={handleEditScoreValue} />,
+      <GameCard isCurrent={false} scoreType="sp" scoreValue="8" onSelectCurrentCard={handleSelectCurrentCard} />,
     );
     const button = getByRole('button');
 
     userEvent.click(button);
-    expect(handleEditScoreValue).toHaveBeenCalledTimes(1);
+    expect(handleSelectCurrentCard).toHaveBeenCalledTimes(1);
 
     userEvent.click(button);
-    expect(handleEditScoreValue).toHaveBeenCalledTimes(2);
+    expect(handleSelectCurrentCard).toHaveBeenCalledTimes(2);
   });
 });
