+++ pointing-poker/src/components/Game/GameCard/index.ts
