+++ pointing-poker/src/components/Game/PlayingCard/index.ts
@@ -1,0 +1,1 @@
+export { PlayingCard } from './PlayingCard';
