+++ pointing-poker/src/components/Game/index.ts
@@ -1,0 +1,2 @@
+export { GameCard } from './GameCard';
+export { GameCardsList } from './GameCardsList';
