--- pointing-poker/src/components/GameCard/GameCard.module.scss
@@ -1,58 +1,0 @@
-@import '../../layout/styles-config';
-
-.gameCard {
-  position: relative;
-  width: 100px;
-  height: 160px;
-  @include flex($direction: column);
-  box-shadow: $card-shadow;
-  border-radius: 5px;
-  margin-bottom: 5px;
-  span {
-    font-size: 14px;
-  }
-  &.active {
-    background-color: #60dabf;
-  }
-}
-.top {
-  width: 90%;
-  @include flex($justify: flex-start);
-}
-.btnEdit {
-  width: 15px;
-  height: 15px;
-  background-image: url(../../assets/icons/icon__edit.svg);
-}
-.center {
-  width: 50px;
-  height: 50px;
-  font-size: 35px;
-  text-transform: uppercase;
-}
-.cardImg {
-  width: 100%;
-  height: 100%;
-  background: url(../../assets/images/img__coffe.png) center no-repeat;
-  background-size: contain;
-}
-.bottom {
-  width: 90%;
-  @include flex($justify: flex-end);
-  span {
-    transform: rotate(180dg);
-  }
-}
-.cover {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 4;
-  &:hover {
-    cursor: pointer;
-    background-color: #60dabf;
-    opacity: 0.5;
-  }
-}
